feat(backoffice-v2): added adapters, connected missing ui to real data
