+++ apps/backoffice-v2/src/common/utils/boolean-to-yes-or-no/boolean-to-yes-or-no.ts
@@ -1,0 +1,7 @@
+export const booleanToYesOrNo = (value: boolean | null | undefined) => {
+  if (typeof value !== 'boolean') {
+    return value;
+  }
+
+  return value ? 'Yes' : 'No';
+};
