+++ apps/backoffice-v2/src/common/utils/check-is-url/check-is-url.ts
@@ -1,0 +1,4 @@
+import { isType } from '@ballerine/common';
+import { z } from 'zod';
+
+export const checkIsUrl = isType(z.string().url());
